requirements.txt
arrow==0.12.1
asn1crypto==0.24.0
backports.shutil-get-terminal-size==1.0.0
backports.shutil-which==3.5.1
binaryornot==0.4.4
bleach==2.1.3
certifi==2018.1.18
cffi==1.11.5
chardet==3.0.4
click==6.7
colorama==0.3.9
cookiecutter==1.6.0
cryptography==2.1.4
cycler==0.10.0
decorator==4.2.1
dominate==2.3.1
entrypoints==0.2.3
Flask==0.12.2
Flask-Bootstrap==3.3.7.1
Flask-Cors==3.0.3
flask-googlemaps==0.2.5
Flask-Login==0.4.0
Flask-SocketIO==2.9.2
Flask-WTF==0.14.2
future==0.16.0
html5lib==1.0.1
idna==2.6
ipykernel==4.8.2
ipython==5.4.1
ipython-genutils==0.2.0
ipywidgets==7.1.2
itsdangerous==0.24
jedi==0.11.1
Jinja2==2.10
jinja2-time==0.2.0
jsonschema==2.6.0
jupyter==1.0.0
jupyter-client==5.2.2
jupyter-console==5.2.0
jupyter-core==4.4.0
jupyter-dashboards-bundlers==0.9.1
kiwisolver==1.0.1
-e git+git@github.com:harshrpg/COMP30670_Assignment3.git@814d5b9187b0820d7ae4dd8be28fc8b021f94544#egg=led_solve
-e git+git@github.com:harshrpg/COMP30670_LedTester.git@ff9bdd0b70ccfa8d3d567a003ee334cd7f31b653#egg=led_tester
MarkupSafe==1.0
matplotlib==2.2.2
mistune==0.8.3
mysql-connector-python==2.0.4
mysql-connector-python-rf==2.2.2
nbconvert==5.3.1
nbformat==4.4.0
notebook==5.4.0
numpy==1.14.1
pandas==0.22.0
pandocfilters==1.4.2
parso==0.1.1
pathlib2==2.3.0
patsy==0.5.0
pickleshare==0.7.4
poyo==0.4.1
prompt-toolkit==1.0.15
pycparser==2.18
Pygments==2.2.0
pyOpenSSL==17.5.0
pyparsing==2.2.0
PySocks==1.6.8
python-dateutil==2.7.0
python-engineio==2.0.1
python-socketio==1.8.3
pytz==2018.3
pywinpty==0.5.1
pyzmq==17.0.0
qtconsole==4.3.1
requests==2.18.4
scandir==1.5
scikit-learn==0.19.1
scipy==1.0.1
seaborn==0.8.1
Send2Trash==1.5.0
simplegeneric==0.8.1
singledispatch==3.4.0.3
six==1.11.0
sklearn==0.0
systeminfo==0.1.1
systeminfo-flask==0.1.0
terminado==0.8.1
testpath==0.3.1
tornado==5.0
traitlets==4.3.2
urllib3==1.22
virtualenv==15.2.0
visitor==0.1.3
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.14.1
whichcraft==0.4.1
widgetsnbextension==3.1.4
win-inet-pton==1.0.1
win-unicode-console==0.5
wincertstore==0.2
WTForms==2.1
